基础部分1-4（？）

+++ src/main/java/com/example/projecthelper/entity/Comment.java
@@ -7,18 +7,19 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Comment {
-    private final Integer comment_id;
-    private final Integer comment_group;
+    private final long comment_id;
+    private final long comment_group;
 
-    private final Integer commented_group;
+    private final long commented_group;
 
-    private final Integer grade;
+    private final long grade;
 
     private final Timestamp comment_time;
 
     private final String content;
 
-    public Comment(Integer comment_id, Integer comment_group, Integer commented_group, Integer grade, Timestamp comment_time, String content) {
+    public Comment(long comment_id, long comment_group, long commented_group,
+                   long grade, Timestamp comment_time, String content) {
         this.comment_id = comment_id;
         this.comment_group = comment_group;
         this.commented_group = commented_group;
