+++ src/main/java/com/example/projecthelper/entity/Notice.java
@@ -1,23 +1,23 @@
 package com.example.projecthelper.entity;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class Notice {
-    private final Integer notice_id;
+    private final long notice_id;
 
     private String title;
     private String content;
-    private final Integer creator_id;
+    private final long creator_id;
 
-    public Notice(Integer notice_id, String title, String content, Integer creator_id) {
+    public Notice(long notice_id, String title, String content, long creator_id) {
         this.notice_id = notice_id;
         this.title = title;
         this.content = content;
         this.creator_id = creator_id;
     }
+
 
 }
